Refactor code to colocate hooks with the page its used in + add "server-only" package to ensure certain code doesn't run on the client (e.g. db queries)
